Only run notification deletion if we actually have some

+++ app/src/main/kotlin/me/proxer/app/fragment/ucp/NotificationFragment.kt
@@ -103,7 +103,7 @@ class NotificationFragment : PagedLoadingFragment<ProxerCall<List<ProxerNotifica
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         when (item.itemId) {
             R.id.delete_all -> {
-                if (!task.isWorking) {
+                if (!innerAdapter.isEmpty() && !task.isWorking) {
                     removalQueue.clear()
                     removalTask.freshExecute(api.notifications().deleteAllNotifications().build())
                 }
